+++ src/gui/Interface.h
@@ -201,7 +201,7 @@ extern Vec2s MemoMouse;
 
 extern long SpecialCursor;
 
-extern long lSLID_VALUE;
+extern float lSLID_VALUE;
 extern bool bInventoryClosing;
 extern E_ARX_STATE_MOUSE eMouseState;
 extern bool bInverseInventory;
